querycache,prep: move PrepareMetadata to new prep package
This is a mechanical change that moves `querycache.PrepareMetadata` to a
new `prep` package in preparation for future changes.

Release note: None
