+++ pkg/sql/plan_opt.go
@@ -133,8 +133,8 @@ func (p *planner) prepareUsingOptimizer(
 
 	if opc.useCache {
 		cachedData, ok := p.execCfg.QueryCache.Find(&p.queryCacheSession, stmt.SQL)
-		if ok && cachedData.PrepareMetadata != nil {
-			pm := cachedData.PrepareMetadata
+		if ok && cachedData.Metadata != nil {
+			pm := cachedData.Metadata
 			// Check that the type hints match (the type hints affect type checking).
 			if !pm.TypeHints.Identical(p.semaCtx.Placeholders.TypeHints) {
 				opc.log(ctx, "query cache hit but type hints don't match")
@@ -229,17 +229,17 @@ func (p *planner) prepareUsingOptimizer(
 			stmt.Prepared.BaseMemo = memo
 		}
 		if opc.useCache {
-			// execPrepare sets the PrepareMetadata.InferredTypes field after this
-			// point. However, once the PrepareMetadata goes into the cache, it
+			// execPrepare sets the Metadata.InferredTypes field after this
+			// point. However, once the Metadata goes into the cache, it
 			// can't be modified without causing race conditions. So make a copy of
 			// it now.
 			// TODO(radu): Determine if the extra object allocation is really
 			// necessary.
-			pm := stmt.Prepared.PrepareMetadata
+			pm := stmt.Prepared.Metadata
 			cachedData := querycache.CachedData{
-				SQL:             stmt.SQL,
-				Memo:            memo,
-				PrepareMetadata: &pm,
+				SQL:      stmt.SQL,
+				Memo:     memo,
+				Metadata: &pm,
 			}
 			p.execCfg.QueryCache.Add(&p.queryCacheSession, &cachedData)
 		}
@@ -821,9 +821,9 @@ func (opc *optPlanningCtx) buildExecMemo(ctx context.Context) (_ *memo.Memo, _ e
 				if err != nil {
 					return nil, err
 				}
-				// Update the plan in the cache. If the cache entry had PrepareMetadata
+				// Update the plan in the cache. If the cache entry had Metadata
 				// populated, it may no longer be valid.
-				cachedData.PrepareMetadata = nil
+				cachedData.Metadata = nil
 				p.execCfg.QueryCache.Add(&p.queryCacheSession, &cachedData)
 				opc.flags.Set(planFlagOptCacheMiss)
 			} else {
