+++ pkg/sql/prep/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "prep",
+    srcs = ["metadata.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/prep",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/sql/catalog/colinfo",
+        "//pkg/sql/parser/statements",
+        "//pkg/sql/sem/tree",
+        "//pkg/sql/types",
+        "@com_github_lib_pq//oid",
+    ],
+)
