+++ pkg/sql/querycache/BUILD.bazel
@@ -2,21 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "querycache",
-    srcs = [
-        "prepared_statement.go",
-        "query_cache.go",
-    ],
+    srcs = ["query_cache.go"],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/querycache",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/sql/catalog/colinfo",
         "//pkg/sql/opt/memo",
-        "//pkg/sql/parser/statements",
-        "//pkg/sql/sem/tree",
-        "//pkg/sql/types",
+        "//pkg/sql/prep",
         "//pkg/util/syncutil",
         "@com_github_cockroachdb_errors//:errors",
-        "@com_github_lib_pq//oid",
     ],
 )
 
@@ -27,6 +20,7 @@ go_test(
     embed = [":querycache"],
     deps = [
         "//pkg/sql/opt/memo",
+        "//pkg/sql/prep",
         "//pkg/util/randutil",
         "@com_github_cockroachdb_errors//:errors",
     ],
